+++ scripts/mounts/mounts.js
@@ -3,9 +3,10 @@ import { sdndConstants } from "../constants.js";
 import { tokens } from "../tokens.js";
 import { utility } from "../utility/utility.js";
 import { getAverageHpFormula } from "../actors/actors.js";
-import { backpacks } from "../backpacks/backpacks.js";
+import { backpacks, gmDropBackpack } from "../backpacks/backpacks.js";
 import { sdndSettings } from "../settings.js";
 import { lockActor } from "../backpacks/backpacks.js";
+import { socket } from "../module.js";
 
 const MOUNTED_EFFECT = {
     "name": "Mounted",
@@ -87,8 +88,14 @@ export let mounts = {
     },
     "changeHealth": async function changeHealth(actorUuid, healthPercentage) {
         let health = (Math.ceil(healthPercentage * 100 / 5) * 5);
+        if (health < 0) {
+            health = 0;
+        }
         let actor = await fromUuid(actorUuid);
         let horse = getHorse(actor);
+        if (!horse) {
+            return;
+        }
         let effect = horse.effects.find(e => e.getFlag(sdndConstants.MODULE_ID, "effectName") == "Mounted");
         if (!effect) {
             return false;
@@ -189,27 +196,58 @@ export let mounts = {
         return true;
     },
     "hooks": {
-        "ready": async function _ready() {
-            if (!sdndSettings.EnableHorseDamage.getValue()) {
-                return false;
-            }
-            if (game.modules.get("midi-qol")) {
-                Hooks.on("midi-qol.AttackRollComplete", onAttackRollComplete);
-                // Hooks.on("midi-qol.DamageRollComplete", onDamageRollComplete);
-            }
-            else {
-                ui.notifications.warn("SDND Mount Damage Requires Midi-QOL!");
-            }
-        },
-        "onDamageTaken": onDamageTaken 
+        "onDamageTaken": onDamageTaken
     },
     "getMountData": getMountData,
-    "setMountData": setMountData
+    "setMountData": setMountData,
+    "getHorse": getHorse
 };
 
 async function onDamageTaken(actor, changes, update, userId) {
     let horse = getHorse(actor);
     if (!horse) {
+        return;
+    }
+    let actorTotalDamage = Math.floor(changes.total / 2);
+    let horseDamage = changes.total - actorTotalDamage;
+    let actorHpDiff = changes.hp - actorTotalDamage;
+    if (actorHpDiff > 0) {
+        changes.hp = 0;
+        changes.temp += actorHpDiff;
+    }
+    else {
+        changes.hp = actorHpDiff;
+    }
+    await applyDamageToHorse(actor, horse, horseDamage);
+}
+
+async function applyDamageToHorse(actor, horse, damage) {
+    let mountData = getMountData(horse);
+    mountData.hp.value += damage;
+    if (mountData.hp.value < 0) {
+        mountData.hp.value = 0;
+    }
+    const isDead = mountData.hp.value <= 0;
+    await setMountData(horse, mountData);
+    await ChatMessage.create({
+        flavor: "Mount injured!",
+        content: `${horse.name} has taken ${damage} hp of damage ${(isDead ? " and has perished." : "")}`
+    });
+    await mounts.changeHealth(actor.uuid, (mountData.hp.value / mountData.hp.max));
+    if (isDead) {
+        await lockActor(actor);
+        let tokenId = actor.getActiveTokens()?.pop()?.id;
+        let userUuid = game.users.find(u => u.character?.uuid == actor.uuid)?.id ?? game.user.id;
+        await gmDropBackpack(tokenId, horse.id, userUuid, true).then(
+            function () {
+                const userId = userUuid?.split(".")?.pop();
+                if (!userId || !tokenId) {
+                    return;
+                }
+                socket.executeForUsers("selectToken", [userId], tokenId);
+            },
+            function (err) { console.log(err.message); }
+        );
         return;
     }
 }
@@ -296,7 +334,7 @@ async function toggleMount() {
         // find the horse
         let horseToken = findHorseToken(actor);
         if (!horseToken) {
-            ui.notifications.warn(`${actor.name} has no owned mount in this scene!`);
+            ui.notifications.warn(`${actor.name} has no living mount in this scene!`);
             return false;
         }
         const distance = tokens.getDistance(controlledToken, horseToken);
@@ -314,7 +352,7 @@ async function toggleMount() {
 }
 
 function getHorse(actor) {
-    let horses = actor.items?.filter(i => i.type == "container" && (i.name?.toLowerCase().includes("horse") || i.img?.toLowerCase()?.includes("horse") ||
+    let horses = actor.items?.filter(i => i.getFlag(sdndConstants.MODULE_ID, "IsMount") || i.type == "container" && (i.name?.toLowerCase().includes("horse") || i.img?.toLowerCase()?.includes("horse") ||
         i.system?.description?.value?.toLowerCase()?.includes("horse")));
     if (horses.length == 1) {
         return horses[0];
@@ -336,7 +374,7 @@ function getAc(item) {
 
 function findHorseToken(actor) {
     return canvas.scene.tokens.find(t => t.actor?.getFlag(sdndConstants.MODULE_ID, "DroppedBy") == actor.uuid &&
-        t.actor?.getFlag(sdndConstants.MODULE_ID, "IsMount"));
+        t.actor?.getFlag(sdndConstants.MODULE_ID, "IsMount") && (t.actor?.system?.attributes?.hp?.value ?? 0) > 0);
 }
 
 async function dismount(token, horse) {
@@ -344,55 +382,3 @@ async function dismount(token, horse) {
     await gmFunctions.dropBackpack(token.id, horse.id, game.user.uuid, true);
 }
 
-async function onAttackRollComplete(workflow) {
-    if (!workflow) {
-        return;
-    }
-    return;
-    // check for missed actors to see if horse is hit
-    for (let target of workflow.targets.filter(t => !workflow.hitTargets.has(t))) {
-        let horse = getHorse(target?.actor);
-        if (!horse) {
-            continue;
-        }
-        let mountData = getMountData(horse);
-        if (mountData.ac.value < workflow.attackTotal) {
-            target["mountHit"] = true;
-            workflow.hitTargets.add(target);
-        }
-    }
-}
-
-async function onDamageRollComplete(workflow) {
-    if (!workflow) {
-        return true;
-    }
-
-    for (let target of workflow.hitTargets.filter(t => t.mountHit)) {
-        let horse = getHorse(target?.actor);
-        if (!horse) {
-            continue;
-        }
-        let mountData = getMountData(horse);
-    }
-
-    return;
-
-    // apply damage to mount
-    let mountData = getMountData(workflow.mountHit);
-    mountData.hp.value -= workflow.damageTotal;
-    // if (mountData.hp.value < 0) {
-    //     workflow.damageTotal = 0 - mountData.hp.value;
-    // }
-    // else {
-    //     workflow.damageTotal = 0;
-    // }
-    workflow.hitTargets.clear();
-    await setMountData(workflow.mountHit, mountData);
-    await mounts.changeHealth(workflow.mountHit.parent.uuid, (mountData.hp.value / mountData.hp.max))
-    await ChatMessage.create({
-        speaker: { "actor": workflow.mountHit.parent },
-        content: `${workflow.mountHit.name} is struck for ${workflow.damageTotal}${mountData.hp.value <= 0 ? " and killed" : ""}...`
-    });
-    return false;
-}