+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # v12.0.11
+# New Features 
+- Mounts
+  - Mounts take damage in combat.  When you take damage while mounted, half of the damage received will be applied to your mount.
 # Bug Fixes
 - Mounts
   - Fix bug with getMountData.